fix: use 32 bit int and float in `MAAValue`
If we pass a big int (larger than i32::MAX) to core, it will crash.

+++ maa-cli/src/run/preset/mod.rs
@@ -34,7 +34,7 @@ pub fn closedown() -> Result<TaskConfig> {
     Ok(task_config)
 }
 
-pub fn fight(stage: String, medicine: Option<i64>) -> Result<TaskConfig> {
+pub fn fight(stage: String, medicine: Option<i32>) -> Result<TaskConfig> {
     let mut task_config = TaskConfig::new();
 
     let mut params = MAAValue::new();
